requirements.txt
django-textify
django-markitup
django-massmedia